refactor(clawdstrike-policy-event): extract kinds/simulation from edr/receipt/mod.rs
