+++ crates/libs/clawdstrike-policy-event/src/edr/receipt/kinds/mod.rs
@@ -2,8 +2,10 @@ mod detection;
 mod evidence_bundle;
 mod graph_slice;
 mod sensor_state;
+mod simulation;
 
 pub(crate) use detection::*;
 pub(crate) use evidence_bundle::*;
 pub(crate) use graph_slice::*;
 pub(crate) use sensor_state::*;
+pub(crate) use simulation::*;
